Pull host output dir logic out
