+++ src/NServiceBus.Core/Hosting/StartupDiagnostics/EndpointDiagnosticSettingsExtensions.cs
@@ -2,9 +2,14 @@
 {
     using Settings;
 
-    //internal for now
+    /// <summary>
+    /// Provides a API to add startup diagnostics.
+    /// </summary>
     static class EndpointDiagnosticSettingsExtensions
     {
+        /// <summary>
+        /// Adds a section to the startup diagnostics.
+        /// </summary>
         public static void AddStartupDiagnosticsSection(this ReadOnlySettings settings, string sectionName, object section)
         {
             Guard.AgainstNull(nameof(settings), settings);
