+++ src/NServiceBus.Core/Routing/RoutingComponent.cs
@@ -58,8 +58,6 @@ namespace NServiceBus
             {
                 EnableBestPracticeEnforcement(conventions, pipelineSettings);
             }
-
-            settings.AddStartupDiagnosticsSection("Routing", this);
         }
 
         static bool ShouldEnforceBestPractices(ReadOnlySettings settings)
